requirements.txt
langchain-community
langchain-openai
openpyxl
streamlit
python-dotenv